Develop (#51)
* fixes formatting in README

* fixes formatting in README, for real this time

* Feature/#16 (#26)

* milestone is used as version when it exists

* always send milestone name

* minor style change

* #17 adds parameter override code

lacks interface in .jelly code though

* add .jelly files for new field

* #19 - Enum for case status and not reporting results for ignored/skipped test cases. Also better clue for when a suite is removed from the TestRail side. (#20)

Looks good to me! TFTPR

* #21fixes problem with extra params not being retained

eliminates crash bug on failure to connect to TestRail

* except for real this time

* fixes problem of the extraParams field not being saved properly

* Adds the ability to skip creating new test cases if they don't already exist in testrail (#25)

* Add new UI element to toggle auto-creation of testcases in testRails.

* Changed UI element, and added logic to skip adding new tests

* #24 adds client code for getting existing runs

* adds ability to select a test run to post results against

* fixes issues with runs

* update jenkins version to a newer LTS release and fixes a bug with the extra parameters handling

* adds workaround for there being only one project

* #35 junit errors are now translated to test rail failures

* failures, not blocks

* remove duplicates from bad merge

* fixes some typos and renames Error -> JunitError to resolve ambiguities

* makes various changes to improve style and appease linters

* add settings file back
