+++ pom.xml
@@ -121,7 +121,7 @@
     <maven-enforcer-plugin.version>1.0</maven-enforcer-plugin.version>
     <maven-help-plugin.version>2.1.1</maven-help-plugin.version>
     <maven-install-plugin.version>2.3.1</maven-install-plugin.version>
-    <maven-javadoc-plugin.version>2.7</maven-javadoc-plugin.version>
+    <maven-javadoc-plugin.version>3.1.0</maven-javadoc-plugin.version>
     <maven-jar-plugin.version>2.3.1</maven-jar-plugin.version>
     <maven-jetty-plugin.version>6.1.26</maven-jetty-plugin.version>
     <maven-jxr-plugin.version>2.2</maven-jxr-plugin.version>
@@ -140,7 +140,7 @@
     <apt-maven-plugin.version>1.0-alpha-4</apt-maven-plugin.version>
     <axistools-maven-plugin.version>1.4</axistools-maven-plugin.version>
     <buildnumber-maven-plugin.version>1.0-beta-4</buildnumber-maven-plugin.version>
-    <build-helper-maven-plugin.version>1.5</build-helper-maven-plugin.version>
+    <build-helper-maven-plugin.version>1.8</build-helper-maven-plugin.version>
     <cargo-maven2-plugin.version>1.0.5</cargo-maven2-plugin.version>
     <cobertura-maven-plugin.version>2.4</cobertura-maven-plugin.version>
     <exec-maven-plugin.version>1.2</exec-maven-plugin.version>
