+++ settings.xml
@@ -1,32 +1,33 @@
+
 <settings>
-  <pluginGroups>
-    <pluginGroup>org.jenkins-ci.tools</pluginGroup>
-  </pluginGroups>
-  <profiles>
-    <profile>
-      <id>jenkins</id>
-      <activation>
-        <activeByDefault>true</activeByDefault>
-      </activation>
-      <repositories>
-        <repository>
-          <id>repo.jenkins-ci.org</id>
-          <url>http://repo.jenkins-ci.org/public/</url>
-        </repository>
-      </repositories>
-      <pluginRepositories>
-        <pluginRepository>
-          <id>repo.jenkins-ci.org</id>
-          <url>http://repo.jenkins-ci.org/public/</url>
-        </pluginRepository>
-      </pluginRepositories>
-    </profile>
-  </profiles>
-  <mirrors>
-    <mirror>
-      <id>repo.jenkins-ci.org</id>
-      <url>http://repo.jenkins-ci.org/public/</url>
-      <mirrorOf>m.g.o-public</mirrorOf>
-    </mirror>
-  </mirrors>
-</settings>
+    <pluginGroups>
+      <pluginGroup>org.jenkins-ci.tools</pluginGroup>
+    </pluginGroups>
+    <profiles>
+      <profile>
+        <id>jenkins</id>
+        <activation>
+          <activeByDefault>true</activeByDefault>
+        </activation>
+        <repositories>
+          <repository>
+            <id>repo.jenkins-ci.org</id>
+            <url>http://repo.jenkins-ci.org/public/</url>
+          </repository>
+        </repositories>
+        <pluginRepositories>
+          <pluginRepository>
+            <id>repo.jenkins-ci.org</id>
+            <url>http://repo.jenkins-ci.org/public/</url>
+          </pluginRepository>
+        </pluginRepositories>
+      </profile>
+    </profiles>
+    <mirrors>
+      <mirror>
+        <id>repo.jenkins-ci.org</id>
+        <url>http://repo.jenkins-ci.org/public/</url>
+        <mirrorOf>m.g.o-public</mirrorOf>
+      </mirror>
+    </mirrors>
+  </settings>
