+++ src/main/java/org/jenkinsci/plugins/testrail/TestRail/Project.java
@@ -16,31 +16,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jenkinsci.plugins.testrail.TestRailObjects;
+package org.jenkinsci.plugins.testrail.TestRail;
 
 /**
  * Created by Drew on 3/24/2014.
  */
-public class Suite {
-    private int id;
-    private String name;
+public class Project {
+    private int _id;
+    private String _name;
 
     public void setId(int id) {
-        this.id = id;
+        _id = id;
     }
 
     public void setName(String name) {
-        this.name = name.trim();
+        _name = name.trim();
     }
 
     public int getId() {
-        return this.id;
+        return _id;
     }
 
     public String getName() {
-        return this.name;
+        return _name;
     }
 
-    public String getStringId() { return Integer.toString(id); }
-
+    public String getStringId() { return Integer.toString(_id); }
 }
