+++ src/main/java/org/jenkinsci/plugins/testrail/TestRail/Suite.java
@@ -16,32 +16,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jenkinsci.plugins.testrail.TestRailObjects;
+package org.jenkinsci.plugins.testrail.TestRail;
 
 /**
  * Created by Drew on 3/24/2014.
  */
-public class Section {
+public class Suite {
     private int id;
-    private int suiteId;
     private String name;
-    private String parentId;
 
     public void setId(int id) {
         this.id = id;
     }
-    public void  setSuiteId(int suiteId) { this.suiteId = suiteId; }
+
     public void setName(String name) {
         this.name = name.trim();
     }
-    public void setParentId(String id) { this.parentId = id; }
 
     public int getId() {
         return this.id;
     }
-    public int getSuiteId() { return this.suiteId; }
+
     public String getName() {
         return this.name;
     }
-    public String getParentId() { return this.parentId; }
+
+    public String getStringId() { return Integer.toString(id); }
+
 }
