+++ src/main/resources/org/jenkinsci/plugins/testrail/TestRailNotifier/help-extraParams.html
@@ -1,0 +1,4 @@
+<div>
+	You can override or include extra parameters for the TestRail report here. Add key value pairs in JSON format.
+	e.g. <span font-family="Lucida Console">{ "version":"1.3b", "defects":"ITD-101, ITD-102", "assignedto_id":"12" }</span>
+</div>
