+++ src/main/resources/org/jenkinsci/plugins/testrail/TestRailNotifier/help-useExistingRun.html
@@ -1,0 +1,4 @@
+<div>
+	You may choose to post your results to an existing Test Run in TestRail, rather than create a new one for each build.
+	(Should you choose to post results to an existing Test Run, the run will *not* be automatically closed by this plugin!)
+</div>
